backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gTTS==2.3.2
nltk==3.8.1
pydub==0.25.1
librosa==0.10.1
tensorflow>=2.16.1
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.2
soundfile==0.12.1 